Export published and finished responses
We were fooled by the tests only covering the happy-path and we missed
that `#unscoped` removed all WHERE constraints of the PublishedQuestions
query.

It turns out that things where a bit more complex and there are other
states of the consultation and questions that need to be considered. The
consultation needs to be finished and have its results published. Then,
questions need to be published as well.