Check question permissions with grantee only
The only person actually present in the assembly and voting is the
grantee. So, if that person is verified so are the votes they place on
behalf of others.

This is done by decoupling permissions of a delegated vote from
a regular one. The former needs to check information from the delegation
whereas, the latter doesn't.

We also ensure the grantee is who performs the vote/unvote. To do that
it was simpler to consider the permissions `user` the one performing the
action (current_user) and simply deal with the delegation's
grantee/granter to check the permissions. This saves us from having to
pass in the `user` argument on the calls to `allowed_to?`.