Fix pagination and remove filtering
Pages is the only thing we care about right now. Filters is a nice to
have but it can come later. This way we reduce the changes of
introducing bugs and we take the time to properly understand how
filtering works. Meanwhile, users (and the serer) can already benefit from pagination.