Fix admin delegations pagination (#77)
* Fix pagination and remove filtering

Pages is the only thing we care about right now. Filters is a nice to
have but it can come later. This way we reduce the changes of
introducing bugs and we take the time to properly understand how
filtering works. Meanwhile, users (and the serer) can already benefit from pagination.

* Remove unused Admin::DelegationPresenter

We stopped using it in b229382 but we didn't remove it. That saves use
because now Kaminari requires us @delegations to be an
ActiveRecord::Relation instead of an Array. We got lucky here.

The underlying Admin::ConsultationPresenter is still used by the
Admin::SettingPresenter.