Display Decidim's results page next to the custom
Instead of making the `Results` menu item point to our custom results
page, we restore that link to point to decidim-consultation's controller
but we link to ours in a new menu item. Uses can now choose which flavor
of results they want.

I had to append decidim-consultation's Permissions class to the chain in
order for the layout to display to two links on the menu. That's because
our permissions class doesn't handle `allowed_to? :read, :question`,
which returned false.