New sum of weights results (#89)
* Add results page summing member weight per answer

* Wrap long lines

* Add new view and query

* Use CAST function

I couldn't find a way to turn PostgreSQL's cast syntax into an Arel
Node. There's a `UnaryOperator` class that enables things like `!column`
but nothing like a `SuffixOperator`.

* Extract different #votes_count

This makes it look very similar to SumOfMembershipWeight which will ease
future refactorings.

* Extract common joins into new query

These build the base result set upon we aggregate different things
depending on the context. In an ideal world this could be a database
view.

* Turn long and complex SQL literal into Arel Nodes

This distributes the various responsibilities and complex bits into
smaller more testable pieces.

It's particularly interesting the idea of having the aggregation as
a separate object which could potentially be injected in the constructor
and allow for subtle different result sets based all on the same query
object.

* Enable exporting by fixing permissions

* Refactor consultation results controllers

This makes them a bit simpler and look alike. I'm not sure it's worth
abstracting anything to remove the duplication.

* Rename query object to something more specific

* Add explanatory header to results tables

Hopefully, this will make it easier for users to understand what's
different between these results pages.