Hide new results pages feature under feature flag (#95)
That is, the work done on #88 and #89.

This allows me to deploy this new feature as it is but not release it
yet. It still needs customer validation but in the meantime, we need to
keep working on other improvements and we don't want to have long-lived
branches. It never ends well.

Also, this won't interfere with other organizations using this module.